astro_app/backend/requirements.txt
fastapi
uvicorn
pyswisseph
pytz
pydantic
pydantic[email]
pydantic-settings
httpx
python-dotenv
sqlalchemy
alembic
passlib[argon2]
python-jose[cryptography]
python-multipart
email-validator
openai
google-generativeai
langchain
langchain-community
langchain-huggingface
chromadb
pypdf
yfinance
pandas
geopy
timezonefinder
requests
